Extract workspace asset workflow hook
